refactoring front-end functionalities
