+++ potato/server_utils/schemas/pure_display.py
@@ -1,0 +1,9 @@
+"""
+Pure Display Layout
+"""
+
+
+def generate_pure_display_layout(annotation_scheme):
+    schematic = '<Strong>%s</Strong> %s' % (annotation_scheme['description'], '<br>'.join(annotation_scheme['labels']))
+
+    return schematic, None
